refactor(cars): Cars 객체에게 메시지를 전달하는 방식으로 변경
