+++ src/main/java/racingcar/domain/Cars.java
@@ -1,6 +1,8 @@
 package racingcar.domain;
 
+import racingcar.constant.Digit;
 import racingcar.constant.Message;
+import racingcar.utils.RandomUtils;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -36,5 +38,34 @@ public class Cars {
 
     public List<Car> getCars() {
         return Collections.unmodifiableList(cars);
+    }
+
+    public void decideMovableCar(List<Integer> randoms) {
+        if (Collections.EMPTY_LIST.equals(randoms)) {
+            randoms = getRandoms(cars);
+        }
+
+        for (int i = 0; i < cars.size(); i++) {
+            moveIfPossible(cars.get(i), randoms.get(i));
+        }
+    }
+
+    private List<Integer> getRandoms(List<Car> cars) {
+        List<Integer> randoms = new ArrayList<>();
+
+        for (int i = 0; i < cars.size(); i++) {
+            int random = RandomUtils.nextInt(Digit.MINIMUM_RANDOM_VALUE.getDigit(),
+                    Digit.MAXIMUM_RANDOM_VALUE.getDigit());
+
+            randoms.add(random);
+        }
+
+        return randoms;
+    }
+
+    public void moveIfPossible(Car car, int random) {
+        if (random >= Digit.MOVEMENT_CRITERION.getDigit()) {
+            car.move();
+        }
     }
 }
